requirements.txt
# leaf requierements
absl-py>=0.10.0
gin-config>=0.4.0
pydub>=0.24.0
lingvo>=0.9.0
numpy== 1.21
tensorflow-addons>=0.12.0
tensorflow-datasets>=4.2.0

# audio 
librosa==0.8.1

#gammatone dependencies 
scipy
nose
mock
matplotlib

# external requirements
torch==1.10.1+cu113 
torchaudio == 0.10.1
soundfile
pandas
PyYAML
# backwards compatibility
# pathlib2
